微信集成之用户管理修改：标签信息新增、编辑时没有任何操作点击关闭不要提示警告信息，全量获取用户按钮放在用户管理页面等

+++ src/main/java/com/tranzvision/gd/TZWeChatUserBundle/service/impl/TzWeChatTagServiceImpl.java
@@ -114,7 +114,7 @@ public class TzWeChatTagServiceImpl extends FrameworkImpl {
 			//当前登录人
 			String oprid = tzLoginServiceImpl.getLoginedManagerOprid(request);
 			
-			String errcodeUnTag = "", errmsgUnTag = "", errcodeTag = "", errmsgTag = "";
+			String errcodeUnTag = "0", errmsgUnTag = "ok", errcodeTag = "0", errmsgTag = "ok";
 
 			int num = 0;
 			for (num = 0; num < actData.length; num++) {
@@ -146,53 +146,56 @@ public class TzWeChatTagServiceImpl extends FrameworkImpl {
 						
 				
 				for(String NoSelectTagId :  NoSelectTagIdArr) {
-					//删除用户标签关系表
-					String sql = "DELETE FROM PS_TZ_WXUSER_TAG_T WHERE TZ_JG_ID=? AND TZ_WX_APPID=? AND TZ_OPEN_ID IN (" + openIdWhere + ") AND TZ_TAG_ID=?";
-					sqlQuery.update(sql,new Object[]{jgId,wxAppId,NoSelectTagId});
-					
-					/*调用微信接口*/
-					
-					/*为用户取消标签*/
-					Map<String, Object> mapUnTag = tzWxApiObject.batchUnTagging(jgId, wxAppId, NoSelectTagId, openIdListUnTag);
-					errcodeUnTag = mapUnTag.get("errcode") == null ? "-1" : mapUnTag.get("errcode").toString();
-					errmsgUnTag = mapUnTag.get("errmsg") == null ? "发生错误，请与系统管理员联系。" : mapUnTag.get("errmsg").toString();
+					if(!"".equals(NoSelectTagId)) {
+						//删除用户标签关系表
+						String sql = "DELETE FROM PS_TZ_WXUSER_TAG_T WHERE TZ_JG_ID=? AND TZ_WX_APPID=? AND TZ_OPEN_ID IN (" + openIdWhere + ") AND TZ_TAG_ID=?";
+						sqlQuery.update(sql,new Object[]{jgId,wxAppId,NoSelectTagId});
+						
+						/*调用微信接口*/
+						
+						/*为用户取消标签*/
+						Map<String, Object> mapUnTag = tzWxApiObject.batchUnTagging(jgId, wxAppId, NoSelectTagId, openIdListUnTag);
+						errcodeUnTag = mapUnTag.get("errcode") == null ? "-1" : mapUnTag.get("errcode").toString();
+						errmsgUnTag = mapUnTag.get("errmsg") == null ? "发生错误，请与系统管理员联系。" : mapUnTag.get("errmsg").toString();
+					}
 				}
 				
 				for(String selectTagId : selectTagIdArr) {
+					if(!"".equals(selectTagId)) {
+						List openIdListTag = new ArrayList<>();
 					
-					List openIdListTag = new ArrayList<>();
-				
-					for(String openId : openIdArr) {	
-					
-						PsTzWxuserTagTKey psTzWxuserTagTKey= new PsTzWxuserTagTKey();
-						psTzWxuserTagTKey.setTzJgId(jgId);
-						psTzWxuserTagTKey.setTzWxAppid(wxAppId);
-						psTzWxuserTagTKey.setTzOpenId(openId);
-						psTzWxuserTagTKey.setTzTagId(selectTagId); 
+						for(String openId : openIdArr) {	
 						
-						PsTzWxuserTagT psTzWxuserTagT = psTzWxuserTagTMapper.selectByPrimaryKey(psTzWxuserTagTKey);
-						if(psTzWxuserTagT==null) {
-							psTzWxuserTagT = new PsTzWxuserTagT();
-							psTzWxuserTagT.setTzJgId(jgId);
-							psTzWxuserTagT.setTzWxAppid(wxAppId);
-							psTzWxuserTagT.setTzOpenId(openId);
-							psTzWxuserTagT.setTzTagId(selectTagId);
-							psTzWxuserTagT.setRowAddedDttm(new Date());
-							psTzWxuserTagT.setRowAddedOprid(oprid);
-							psTzWxuserTagT.setRowLastmantDttm(new Date());
-							psTzWxuserTagT.setRowLastmantOprid(oprid);
-							psTzWxuserTagTMapper.insertSelective(psTzWxuserTagT);
+							PsTzWxuserTagTKey psTzWxuserTagTKey= new PsTzWxuserTagTKey();
+							psTzWxuserTagTKey.setTzJgId(jgId);
+							psTzWxuserTagTKey.setTzWxAppid(wxAppId);
+							psTzWxuserTagTKey.setTzOpenId(openId);
+							psTzWxuserTagTKey.setTzTagId(selectTagId); 
 							
-							openIdListTag.add(openId);
+							PsTzWxuserTagT psTzWxuserTagT = psTzWxuserTagTMapper.selectByPrimaryKey(psTzWxuserTagTKey);
+							if(psTzWxuserTagT==null) {
+								psTzWxuserTagT = new PsTzWxuserTagT();
+								psTzWxuserTagT.setTzJgId(jgId);
+								psTzWxuserTagT.setTzWxAppid(wxAppId);
+								psTzWxuserTagT.setTzOpenId(openId);
+								psTzWxuserTagT.setTzTagId(selectTagId);
+								psTzWxuserTagT.setRowAddedDttm(new Date());
+								psTzWxuserTagT.setRowAddedOprid(oprid);
+								psTzWxuserTagT.setRowLastmantDttm(new Date());
+								psTzWxuserTagT.setRowLastmantOprid(oprid);
+								psTzWxuserTagTMapper.insertSelective(psTzWxuserTagT);
+								
+								openIdListTag.add(openId);
+							}
 						}
+					
+						/*调用微信接口*/
+						
+						/*为用户打标签*/
+						Map<String, Object> mapTag = tzWxApiObject.batchTagging(jgId, wxAppId, selectTagId, openIdListTag);
+						errcodeTag = mapTag.get("errcode") == null ? "-1" : mapTag.get("errcode").toString();
+						errmsgTag = mapTag.get("errmsg") == null ? "发生错误，请与系统管理员联系。" : mapTag.get("errmsg").toString();
 					}
-					
-					/*调用微信接口*/
-					
-					/*为用户打标签*/
-					Map<String, Object> mapTag = tzWxApiObject.batchTagging(jgId, wxAppId, selectTagId, openIdListTag);
-					errcodeTag = mapTag.get("errcode") == null ? "-1" : mapTag.get("errcode").toString();
-					errmsgTag = mapTag.get("errmsg") == null ? "发生错误，请与系统管理员联系。" : mapTag.get("errmsg").toString();
 				}
 				
 				mapRet.put("errcodeUnTag", errcodeUnTag);
